Feature/confirmation dialog

+++ src/app/components/workspaces/workspace-list-item/workspace-list-item.component.ts
@@ -3,8 +3,10 @@ import { Router } from '@angular/router';
 import { ProductService } from 'src/app/services/api/product.service';
 import { WorkspaceService } from 'src/app/services/api/workspace.service';
 import { ConstantsService } from 'src/app/services/constants.service';
+import { ConfirmationDialogComponent, ConfirmationDialogModel } from 'src/app/shared/dialogs/confirmation-dialog/confirmation-dialog.component';
 import { Workspace } from 'src/app/shared/models/workspace.model';
 import { WorkspaceViewModel } from '../workspaces.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-workspace-list-item',
@@ -16,7 +18,7 @@ export class WorkspaceListItemComponent {
   @Output() activeWorkspace = new EventEmitter<Workspace>();
   @Output() deletedWorkspace = new EventEmitter<Workspace>();
 
-  constructor(private oConstantsService: ConstantsService, private oRouter: Router, private oWorkspaceService: WorkspaceService) { }
+  constructor(private oConstantsService: ConstantsService, public oDialog: MatDialog, private oRouter: Router, private oWorkspaceService: WorkspaceService) { }
 
   getDate(sTimestamp: number) {
     if (sTimestamp) {
@@ -40,27 +42,40 @@ export class WorkspaceListItemComponent {
   }
 
   deleteWorkspace(sWorkspaceId: string) {
+    let sMessage = "Are you sure you wish to delete this Workspace?"
 
-    let oWorkspaceViewModel: WorkspaceViewModel;
-    let oActiveWorkspace: Workspace;
+    let dialogData = new ConfirmationDialogModel("Confirm Deletion", sMessage);
 
-    this.oWorkspaceService.getWorkspaceEditorViewModel(sWorkspaceId).subscribe(oResponse => {
-      oWorkspaceViewModel = oResponse;
-      //add confirmation dialog -> booleans will be inside dialog confirmation 
-      console.log(oWorkspaceViewModel)
-      if (oWorkspaceViewModel) {
-        let bDeleteFile: boolean = true;
-        let bDeleteLayer: boolean = true;
-        this.oWorkspaceService.deleteWorkspace(oWorkspaceViewModel, bDeleteFile, bDeleteLayer)?.subscribe(oResponse => {
-          oActiveWorkspace = this.oConstantsService.getActiveWorkspace();
-          if (JSON.stringify(oActiveWorkspace) === JSON.stringify(oWorkspaceViewModel)) {
-            oWorkspaceViewModel = {} as WorkspaceViewModel;
+    let dialogRef = this.oDialog.open(ConfirmationDialogComponent, {
+      maxWidth: "400px",
+      data: dialogData
+    })
 
-            this.oConstantsService.setActiveWorkspace(oWorkspaceViewModel);
-          }
-          this.deletedWorkspace.emit(oWorkspaceViewModel)
-        })
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult === false) {
+        return false;
       }
+
+      let oWorkspaceViewModel: WorkspaceViewModel;
+      let oActiveWorkspace: Workspace;
+
+      this.oWorkspaceService.getWorkspaceEditorViewModel(sWorkspaceId).subscribe(oResponse => {
+        oWorkspaceViewModel = oResponse;
+
+        if (oWorkspaceViewModel) {
+          let bDeleteFile: boolean = true;
+          let bDeleteLayer: boolean = true;
+          this.oWorkspaceService.deleteWorkspace(oWorkspaceViewModel, bDeleteFile, bDeleteLayer)?.subscribe(oResponse => {
+            oActiveWorkspace = this.oConstantsService.getActiveWorkspace();
+            if (JSON.stringify(oActiveWorkspace) === JSON.stringify(oWorkspaceViewModel)) {
+              oWorkspaceViewModel = {} as WorkspaceViewModel;
+              this.oConstantsService.setActiveWorkspace(oWorkspaceViewModel);
+            }
+            this.deletedWorkspace.emit(oWorkspaceViewModel)
+          })
+        }
+      })
+      return true;
     })
   }
